+++ src/components/implementation/other/vas_test/vas_test.c
@@ -11,7 +11,8 @@ void cos_init(void)
 	printc("vas expand @ %x.\n", (unsigned int)a);
 	if (a == 0) return;
 	t = mman_get_page(cos_spd_id(), a+SERVICE_SIZE, 0);
-	printc("mapped page %x, target %x\n", t, a+SERVICE_SIZE);
+	printc("mapped page %x, target %x\n", 
+	       (unsigned int)t, (unsigned int)(a+SERVICE_SIZE));
 	if (t != a+SERVICE_SIZE) printc("vas ERROR\n");
 	
 	return;
